requirements.txt
Flask==2.3.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
mysql-connector-python==8.0.33
python-dotenv==1.0.0